requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.8.3=pypi_0
aiosignal=1.2.0=pypi_0
async-timeout=4.0.2=pypi_0
attrs=22.1.0=pypi_0
blas=1.0=mkl
bottleneck=1.3.5=py38h080aedc_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotlipy=0.7.0=py38h294d835_1004
ca-certificates=2022.10.11=haa95532_0
certifi=2022.9.24=py38haa95532_0
cffi=1.15.1=py38hd8c33c5_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=pypi_0
colorama=0.4.5=py38haa95532_0
cryptography=37.0.4=py38hb7941b4_0
cudatoolkit=11.6.0=hc0ea762_10
cycler=0.11.0=pyhd3eb1b0_0
dataclasses=0.8=pyh6d0b6a4_7
datasets=2.6.1=pypi_0
dill=0.3.5.1=pypi_0
docker-pycreds=0.4.0=pypi_0
evaluate=0.3.0=pypi_0
fftw=3.3.9=h2bbff1b_1
filelock=3.8.0=pypi_0
flask=2.1.3=py38haa95532_0
flask-cors=3.0.10=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h546665d_0
frozenlist=1.3.1=pypi_0
fsspec=2022.10.0=pypi_0
gitdb=4.0.9=pypi_0
gitpython=3.1.29=pypi_0
glib=2.69.1=h5dc1a3c_1
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
huggingface-hub=0.10.1=pypi_0
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.4=pyhd8ed1ab_0
importlib-metadata=4.11.3=py38haa95532_0
intel-openmp=2021.4.0=haa95532_3556
itsdangerous=2.0.1=pyhd3eb1b0_0
jinja2=3.1.2=py38haa95532_0
joblib=1.2.0=pypi_0
jpeg=9e=h8ffe710_2
kiwisolver=1.4.2=py38hd77b12b_0
lcms2=2.12=h2a16943_0
lerc=3.0=h0e60522_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=12.0.0=default_h627e005_2
libdeflate=1.12=h8ffe710_0
libffi=3.4.2=hd77b12b_4
libiconv=1.16=h2bbff1b_2
libogg=1.3.5=h2bbff1b_1
libpng=1.6.37=h1d00b33_4
libtiff=4.4.0=h2ed3b44_1
libuv=1.44.2=h8ffe710_0
libvorbis=1.3.7=he774522_0
libwebp=1.2.4=h2bbff1b_0
libwebp-base=1.2.4=h8ffe710_0
libxcb=1.13=hcd874cb_1004
libxml2=2.9.14=h0ad7f3c_0
libxslt=1.1.35=h2bbff1b_0
libzlib=1.2.12=h8ffe710_2
lz4-c=1.9.3=h8ffe710_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=2.1.1=py38h2bbff1b_0
matplotlib=3.5.2=py38haa95532_0
matplotlib-base=3.5.2=py38hd77b12b_0
mkl=2021.4.0=haa95532_640
mkl-include=2022.1.0=h6a75c08_874
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.1=py38h277e83a_0
mkl_random=1.2.2=py38hf11a4ad_0
msys2-conda-epoch=20160418=1
multidict=6.0.2=pypi_0
multiprocess=0.70.13=pypi_0
munkres=1.1.4=py_0
nltk=3.7=pyhd3eb1b0_0
numexpr=2.8.4=py38h5b0cc5e_0
numpy=1.23.4=py38h3b20f71_0
numpy-base=1.23.4=py38h4da318b_0
openjpeg=2.5.0=hc9384bd_1
openssl=1.1.1s=h2bbff1b_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.5.1=py38hf11a4ad_0
pathtools=0.1.2=pypi_0
pcre=8.45=hd77b12b_0
pillow=9.2.0=py38h37aa274_2
pip=22.2.2=py38haa95532_0
ply=3.11=py38_0
promise=2.3=pypi_0
protobuf=4.21.8=pypi_0
psutil=5.9.3=pypi_0
pthread-stubs=0.4=hcd874cb_1001
pyarrow=9.0.0=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_1
pyparsing=3.0.9=py38haa95532_0
pyqt=5.15.7=py38hd77b12b_0
pyqt5-sip=12.11.0=py38hd77b12b_0
pysocks=1.7.1=pyh0701188_6
python=3.8.13=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=1.12.1=py3.8_cuda11.6_cudnn8_0
pytorch-mutex=1.0=cuda
pytz=2022.5=pypi_0
pyyaml=6.0=pypi_0
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=hb9a9bb5_4
qtwebkit=5.212=h3ad3cdb_4
regex=2022.9.13=pypi_0
requests=2.28.1=pyhd8ed1ab_1
responses=0.18.0=pypi_0
scikit-learn=1.1.2=pypi_0
scipy=1.9.3=pypi_0
seaborn=0.12.0=py38haa95532_0
sentry-sdk=1.10.1=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=63.4.1=py38haa95532_0
shortuuid=1.0.9=pypi_0
sip=6.6.2=py38hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0=pypi_0
smmap=5.0.0=pypi_0
sqlite=3.39.3=h2bbff1b_0
tbb=2021.5.0=h2d74725_1
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h8ffe710_0
tokenizers=0.13.1=pypi_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=0.12.1=py38_cu116
torchvision=0.13.1=py38_cu116
tornado=6.2=py38h2bbff1b_0
tqdm=4.64.1=py38haa95532_0
transformers=4.23.1=pypi_0
typing_extensions=4.4.0=pyha770c72_0
urllib3=1.26.11=pyhd8ed1ab_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py38haa244fe_4
wincertstore=0.2=py38haa95532_2
xorg-libxau=1.0.9=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xxhash=3.1.0=pypi_0
xz=5.2.6=h8d14728_0
yarl=1.8.1=pypi_0
zipp=3.8.0=py38haa95532_0
zlib=1.2.12=h8cc25b3_2
zstd=1.5.2=h6255e5f_4
